prod/agent/requirements.txt
livekit>=1.0.0
openai>=1.0.0
numpy>=1.20.0
python-dotenv>=0.19.0
pydub>=0.25.1
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
